backend/requirements.txt
# Core framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AWS SDK and services
boto3==1.34.0
botocore==1.34.0

# HTTP client for Meilisearch
aiohttp==3.9.0
asyncio-throttle==1.0.2

# Image processing
Pillow==10.1.0

# Date/time handling
python-dateutil==2.8.2

# Environment variable management
python-dotenv==1.0.0

# Logging and monitoring
structlog==23.2.0

# Testing dependencies (development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2

# Type checking (development)
mypy==1.7.1
types-boto3==1.0.2

# Code formatting and linting (development)
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Security
cryptography==41.0.8

# JSON handling
orjson==3.9.10

# Validation and serialization
marshmallow==3.20.1

# Configuration management
pydantic-settings==2.1.0

# Background task processing
celery==5.3.4
redis==5.0.1

# Database migrations (if using Alembic for DynamoDB schema management)
alembic==1.13.1

# CORS middleware
fastapi-cors==0.0.6

# Authentication and JWT
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Development utilities
pre-commit==3.5.0
bandit==1.7.5
